+++ variables.tf
@@ -133,6 +133,10 @@ variable "ingress_https_nodeport" {
     default = 32443
 }
 
+variable "ingress_https_traffic_enabled" {
+    default = false
+}
+
 variable "ingress_service_monitor_enabled" {
     default = false
 }
@@ -142,6 +146,32 @@ variable "ingress_requests_cpu" {
 }
 
 variable "ingress_requests_memory" {
+    default = "90Mi"
+}
+
+# ================== ingress-nginx-additional =================
+
+variable "helm_ingress_nginx_additional_enabled" {
+    default = false
+}
+
+variable "ingress_additional_http_nodeport" {
+    default = 31080
+}
+
+variable "ingress_additional_https_nodeport" {
+    default = 31443
+}
+
+variable "ingress_additional_https_traffic_enabled" {
+    default = false
+}
+
+variable "ingress_additional_requests_cpu" {
+    default = "100m"
+}
+
+variable "ingress_additional_requests_memory" {
     default = "90Mi"
 }