+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [4.2.0] - 2022-11-10
+
+### Added
+- Additional nginx controller instance
+- Enable https traffic to nginx 
+
 ## [4.1.1] - 2022-11-04
 
 ### Added